Rename “reloadServer” to “server”
